server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
langchain
langchain-openai
langgraph
langsmith
sqlalchemy
alembic
psycopg2-binary
python-multipart
PyYAML
nacos-sdk-python==3.0.4
nacos-sdk-python==3.0.4
